movielens-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
